+++ apps/commerce-api/src/main/java/com/loopers/application/point/PointResult.java
@@ -2,10 +2,12 @@ package com.loopers.application.point;
 
 import com.loopers.domain.point.PointInfo;
 
+import java.math.BigDecimal;
+
 public record PointResult(
         Long id,
         String loginId,
-        Long amount
+        BigDecimal amount
 ) {
     public static PointResult from(final PointInfo pointInfo) {
         return new PointResult(
