+++ apps/commerce-api/src/main/java/com/loopers/domain/payment/PaymentCommand.java
@@ -1,8 +1,10 @@
 package com.loopers.domain.payment;
 
+import java.math.BigDecimal;
+
 public class PaymentCommand {
     public record Payment(
             Long userId,
-            int totalPrice
+            BigDecimal totalPrice
     ) {}
 }
