+++ apps/commerce-api/src/main/java/com/loopers/interfaces/api/point/PointV1Controller.java
@@ -7,6 +7,8 @@ import com.loopers.interfaces.api.ApiResponse;
 import com.loopers.support.error.CoreException;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
+
 @RestController
 @RequestMapping("/api/v1/points")
 public class PointV1Controller implements PointV1ApiSpec {
@@ -32,7 +34,7 @@ public class PointV1Controller implements PointV1ApiSpec {
 
     @PostMapping("/charge")
     @Override
-    public ApiResponse<PointV1Dto.GetResponse> charge(@RequestHeader(value = "X-USER-ID", required = true) String loginId, @RequestBody Long amount) throws CoreException {
+    public ApiResponse<PointV1Dto.GetResponse> charge(@RequestHeader(value = "X-USER-ID", required = true) String loginId, @RequestBody BigDecimal amount) throws CoreException {
 
         PointV1Dto.GetResponse response = PointV1Dto.GetResponse.from(
                 pointFacade.charge(new PointCriteria.Charge(loginId, amount))
